split unsafe jar into smaller, self-contained module
